+++ crates/ruma-client-api/src/media/get_content_thumbnail.rs
@@ -27,11 +27,16 @@ pub mod v3 {
         history: {
             1.0 => "/_matrix/media/r0/thumbnail/:server_name/:media_id",
             1.1 => "/_matrix/media/v3/thumbnail/:server_name/:media_id",
+            1.11 => deprecated,
         }
     };
 
     /// Request type for the `get_content_thumbnail` endpoint.
     #[request(error = crate::Error)]
+    #[deprecated = "\
+        Since Matrix 1.11, clients should use `authenticated_media::get_content_thumbnail::v1::Request` \
+        instead if the homeserver supports it.\
+    "]
     pub struct Request {
         /// The server name from the mxc:// URI (the authoritory component).
         #[ruma_api(path)]
@@ -124,6 +129,7 @@ pub mod v3 {
         pub cross_origin_resource_policy: Option<String>,
     }
 
+    #[allow(deprecated)]
     impl Request {
         /// Creates a new `Request` with the given media ID, server name, desired thumbnail width
         /// and desired thumbnail height.