+++ crates/ruma-client-api/src/media/get_media_preview.rs
@@ -21,11 +21,16 @@ pub mod v3 {
         history: {
             1.0 => "/_matrix/media/r0/preview_url",
             1.1 => "/_matrix/media/v3/preview_url",
+            1.11 => deprecated,
         }
     };
 
     /// Request type for the `get_media_preview` endpoint.
     #[request(error = crate::Error)]
+    #[deprecated = "\
+      Since Matrix 1.11, clients should use `authenticated_media::get_media_preview::v1::Request` \
+      instead if the homeserver supports it.\
+    "]
     pub struct Request {
         /// URL to get a preview of.
         #[ruma_api(query)]
@@ -49,6 +54,7 @@ pub mod v3 {
         pub data: Option<Box<RawJsonValue>>,
     }
 
+    #[allow(deprecated)]
     impl Request {
         /// Creates a new `Request` with the given url.
         pub fn new(url: String) -> Self {